backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.25
aiosqlite==0.19.0

# PostgreSQL (opcional)
psycopg2-binary==2.9.9

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Email
aiosmtplib==3.0.1

# Excel export
openpyxl==3.1.2

# Utilidades
python-dateutil==2.8.2

# Testing (desarrollo)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
